added catch for logo files that are already renamed/exist

+++ unhash.ps1
@@ -48,13 +48,16 @@ for ($i = 0; $i -lt $b.count; $i++) {
     # }
     #pattern = order_design_{4-6 digits}-{fundId}_{logoSize}_{logoScript}
     if ($newName -eq "") { return }
-    if (Test-Path $newName) {
-        Write-Output "`t deleting file $newName`n"
+    if ((Test-Path $newName) -and ($newName -ne $b[$i].name)) {
+        # Write-Output "`t deleting file $newName`n"
+        Write-Output "newName: $newName"
+        Write-Output "b[i]: $b[$i].name"
         # Write-Host -NoNewLine 'Press any key to continue...';
         # $null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown');
         Remove-Item $b[$i]
     }
     else {
+        if ($newName -eq $b[$i].name) { return }
         Write-Output "`t renaming file: $newName`n"
         # Write-Host -NoNewLine 'Press any key to continue...';
         # $null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown');
